* Yet another attempt to import DynamicClient
