+++ requirements.txt
@@ -1,2 +1,3 @@
 openshift>=0.6.2
 requests-oauthlib
+kubernetes
