+++ tests/unit/requirements.txt
@@ -1,2 +1,3 @@
 pytest
 PyYAML
+kubernetes
